Add file to generate document vectors from test and debug getAbstracts
Debug error in getAbstracts that causes question marks to be added to the end of PMIDs that are less than 8 characters long. Fix by changing the line of code that gets the PMID to just take from the 6th character to the last one (exclusive, to strip newline). Unclear why precisely this fixes the error - previously it was assumed that all PMID were 8 characters long, took from 6th to 14th, but PMIDs with 5 and 7 characters only had one question mark added. Error appears to be fixed in this commit, but keep an eye out for weird behavior.