Add functional script to filter down the graph
TODO: There is a silent bug in keyword_cluster_filter() that causes
things not in the cluster to be included, need to resolve this. Also,
there is a bug in PyGraphviz's G.copy() that makes the copy of a directed
graph not be a directed graph (see issue #354 on the pygraphviz repo).